+++ gateway/src/main/java/ru/practicum/shareit/exception/Error.java
@@ -1,0 +1,10 @@
+package ru.practicum.shareit.exception;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class Error {
+    private String error;
+}
