+++ gateway/src/main/java/ru/practicum/shareit/exception/ErrorHandler.java
@@ -1,0 +1,20 @@
+package ru.practicum.shareit.exception;
+
+import jakarta.validation.ConstraintViolationException;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.ResponseStatus;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+
+@RestControllerAdvice
+@Slf4j
+public class ErrorHandler {
+    @ExceptionHandler
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    public Error handleMethodArgumentNotValidException(final ConstraintViolationException e) {
+        log.error("Пользователь передал неверные данные для создания объекта");
+        return new Error(e.getMessage());
+    }
+}
