+++ server/src/main/java/ru/practicum/shareit/ShareItServer.java
@@ -5,7 +5,6 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
 public class ShareItServer {
-
     public static void main(String[] args) {
         SpringApplication.run(ShareItServer.class, args);
     }
